Moved @aragon/apps-* to devDependencies, pre-write the known roles JSON file
