+++ package.json
@@ -8,10 +8,11 @@
     "dao": "./aliases/dao"
   },
   "scripts": {
-    "build": "babel -d dist src",
+    "build": "npm run extract-roles; babel -d dist src --copy-files",
     "prepare": "npm run build",
     "test": "ava",
-    "lint": "standard src"
+    "lint": "standard src",
+    "extract-roles": "./scripts/extractAppRoles.js"
   },
   "repository": {
     "type": "git",
@@ -35,10 +36,6 @@
   "homepage": "https://github.com/aragon/aragon-dev-cli#readme",
   "dependencies": {
     "@aragon/apm": "^2.0.1",
-    "@aragon/apps-finance": "^1.0.1",
-    "@aragon/apps-token-manager": "^1.0.0",
-    "@aragon/apps-vault": "^2.0.1",
-    "@aragon/apps-voting": "^1.0.0",
     "@aragon/aragen": "2.0.0-beta.1",
     "@aragon/os": "4.0.0-beta.1",
     "@aragon/templates-beta": "^1.1.3",
@@ -73,6 +70,10 @@
     "yargs": "^10.1.0"
   },
   "devDependencies": {
+    "@aragon/apps-finance": "^1.0.1",
+    "@aragon/apps-token-manager": "^1.0.0",
+    "@aragon/apps-vault": "^2.0.1",
+    "@aragon/apps-voting": "^1.0.0",
     "@babel/cli": "^7.0.0-beta.46",
     "@babel/core": "^7.0.0-beta.46",
     "@babel/node": "^7.0.0-beta.46",
